+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Optimize the trappings search
 - Manage trapping to be equiped/wearing by default
 - add deutsch compatibility
+- Limit generation button fore those who can create actor
 
 ### Bugfix